refactor: update the default directories
Update the default directory from `/app` to `/`, representing the root.
As subdomains are being used, the additional subdirectory is not necessary
as a default directory.

Improve the regex of filterPathname for label. The regex now correctly
matches any string after `/label`.
