+++ middleware.ts
@@ -7,14 +7,12 @@ export async function middleware(req: NextRequest) {
   switch (req.nextUrl.pathname) {
     case '/auth':
       if (token) {
-        return NextResponse.redirect(new URL('/app', req.url));
+        return NextResponse.redirect(new URL('/', req.url));
       }
       break;
-    case '/':
-      return NextResponse.redirect(new URL('/app', req.url));
   }
 }
 
 export const config = {
-  matcher: ['/', '/auth'],
+  matcher: ['/auth'],
 };
