+++ pages/[...home].tsx
