+++ pages/index.tsx
@@ -1,16 +1,32 @@
-import { LayoutHome } from '@layouts/layoutHome';
-import { ReactElement } from 'react';
+import { ErrorState } from '@components/loadable/errorState';
+import { LoadingTodos } from '@components/loadable/loadingStates/loadingTodos';
+import { LayoutApp } from '@layouts/layoutApp';
+import dynamic from 'next/dynamic';
+import { Fragment as AppFragment, ReactElement, Suspense } from 'react';
 
-const Home = () => {
+const TodoList = dynamic(() => import('components/todos/todoList').then((mod) => mod.TodoList), {
+  ssr: false,
+});
+const FilterTodoIdsEffect = dynamic(() =>
+  import('@states/todos/filterTodoIdsEffect').then((mod) => mod.FilterTodoIdsEffect),
+);
+const ErrorBoundary = dynamic(() => import('react-error-boundary').then((mod) => mod.ErrorBoundary));
+
+const App = () => {
   return (
-    <div className='mt-10 flex w-full flex-row items-center justify-center text-2xl'>
-      This area is intentionally left blank for now.
-    </div>
+    <AppFragment>
+      <ErrorBoundary fallback={<ErrorState />}>
+        <Suspense fallback={<LoadingTodos />}>
+          <FilterTodoIdsEffect />
+          <TodoList />
+        </Suspense>
+      </ErrorBoundary>
+    </AppFragment>
   );
 };
 
-Home.getLayout = function getLayout(page: ReactElement) {
-  return <LayoutHome>{page}</LayoutHome>;
+App.getLayout = function getLayout(page: ReactElement) {
+  return <LayoutApp>{page}</LayoutApp>;
 };
 
-export default Home;
+export default App;
